+++ mesos/handle_offers.go
@@ -35,12 +35,10 @@ func HandleOffers(offers *mesosproto.Event_Offers) error {
 
 	select {
 	case cmd := <-config.CommandChan:
-
 		firstOffer := offers.Offers[0]
 		agentID := offerIds[0].Value
 
 		var taskInfo []*mesosproto.TaskInfo
-		//RefuseSeconds := 5.0
 
 		switch cmd.ContainerType {
 		case "NONE":
@@ -59,9 +57,6 @@ func HandleOffers(offers *mesosproto.Event_Offers) error {
 				OfferIds: []*mesosproto.OfferID{{
 					Value: agentID,
 				}},
-				//Filters: &mesosproto.Filters{
-				//	RefuseSeconds: &RefuseSeconds,
-				//},
 				Operations: []*mesosproto.Offer_Operation{{
 					Type: mesosproto.Offer_Operation_LAUNCH.Enum(),
 					Launch: &mesosproto.Offer_Operation_Launch{
@@ -71,11 +66,19 @@ func HandleOffers(offers *mesosproto.Event_Offers) error {
 		logrus.Debug("Offer Accept: ", offerIds)
 		return Call(accept)
 	default:
+		// decline unneeded offer
 		logrus.Debug("Offer Decline: ", offerIds)
 		decline := &mesosproto.Call{
 			Type:    mesosproto.Call_DECLINE.Enum(),
 			Decline: &mesosproto.Call_Decline{OfferIds: offerIds},
 		}
-		return Call(decline)
+		Call(decline)
+
+		// tell mesos he dont have to offer again until we ask
+		logrus.Debug("Framework Suppress: ", offerIds)
+		suppress := &mesosproto.Call{
+			Type: mesosproto.Call_SUPPRESS.Enum(),
+		}
+		return Call(suppress)
 	}
 }