+++ mesos/mesos.go
@@ -53,7 +53,6 @@ func Subscribe() error {
 		TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
 	}
 
-	logrus.Debug(config.MesosSSL)
 	protocol := "https"
 	if config.MesosSSL == false {
 		protocol = "http"