+++ changelog.md
@@ -17,3 +17,4 @@
 - Add Support to configure (non)SSL Support to Mesos
 - Add Custom Domain to Zookeeper and Kafka to match Consul DNS
 - Add Service Name ENV Variable to Zookeeper and Kafka to match Consul DNS
+- Add Call_Suppress to tell mesos it does not send us offers until we ask